Add optional `date` argument to `builtins.fetchGit`
This allows users to specify an absolute or relative
date (basically, anything that git accepts as a date
specification) when fetching a repository.

The motivation for this change is to enable support
for incremental builds for Haskell packages for this
Nixpkgs branch:

https://github.com/MercuryTechnologies/nixpkgs/tree/gabriella/incremental

… inspired by this blog post:

https://harry.garrood.me/blog/easy-incremental-haskell-ci-builds-with-ghc-9.4/

Keep in mind that this new feature can power incremental
builds for other package managers, too.  There is not
much that is Haskell-specific about this feature.

The basic idea is that instead of Nix doing a full build
for a package, we split every build into two builds:

- A full build at an older point in time

  e.g. a daily or weekly time boundary

- An incremental build relative to the last full build

  This incremental build reuses the build products left
  over from the most recent full build.

In order to do this, though, we need a way to "snap" a
package's `git` source input to an earlier point in time
(e.g. a daily boundary or weekly boundary).  This would
allow multiple incremental builds to share the same
full rebuild if they snap to the same time boundary.

The two main approaches I considered were:

- Approach 1 (this PR)

  Patch Nix to add a `date` argument to `builtins.fetchGit`

- Approach 2

  - Patch `nix-prefetch-git` to support a new `--date` option

  - Disable the sandbox

  - Run `nix-prefetch-git` at evaluation time using
    import-from-derivation to fetch and rehash the repository
    at an older point in time

Approach 1 seemed the more desirable of the two.